requirements.txt
absl-py==2.0.0
aiohttp==3.8.5
aiosignal==1.3.1
anndata==0.9.2
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
arrow==1.3.0
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
boto3==1.28.68
botocore==1.31.68
certifi==2023.7.22
charset-normalizer==3.2.0
chex==0.1.7
click==8.1.7
cmake==3.25.0
comm==0.1.4
contextlib2==21.6.0
contourpy==1.1.1
croniter==1.4.1
cycler==0.11.0
cytoolz==0.12.2
dateutils==0.6.12
debugpy==1.8.0
decorator==5.1.1
deepdiff==6.6.1
dm-tree==0.1.8
docker-pycreds==0.4.0
docopt==0.6.2
docrep==0.3.2
docstring-parser==0.15
einops==0.7.0
etils==1.5.1
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.104.0
filelock==3.9.0
flax==0.7.4
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.38
h11==0.14.0
h5py==3.9.0
idna==3.4
importlib-resources==6.1.0
iniconfig==2.0.0
inquirer==3.1.3
ipykernel==6.25.2
ipython==8.15.0
ipywidgets==8.1.1
itsdangerous==2.1.2
jax==0.4.19
jaxlib==0.4.19
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lightning==2.0.9.post0
lightning-cloud==0.5.43
lightning-utilities==0.9.0
lit==15.0.7
llvmlite==0.41.0
loompy==3.0.7
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.3.1
mpmath==1.3.0
msgpack==1.0.7
mudata==0.2.3
multidict==6.0.4
multipledispatch==1.0.0
natsort==8.4.0
nest-asyncio==1.5.8
networkx==3.0
numba==0.58.0
numpy==1.25.2
numpy-groupies==0.10.2
numpyro==0.13.2
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.1
ordered-set==4.1.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pastel==0.2.1
pathtools==0.1.2
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pipreqs==0.4.13
platformdirs==3.10.0
pluggy==1.3.0
poethepoet==0.10.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.16.1
PyJWT==2.8.0
pynndescent==0.5.10
pyparsing==3.1.1
pyro-api==0.1.2
pyro-ppl==1.8.6
pytest==7.4.2
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.9
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
readchar==4.0.5
requests==2.31.0
rich==13.5.3
s3transfer==0.7.0
scanpy==1.9.5
scikit-learn==1.1.3
scipy==1.11.3
scvi-tools==1.0.4
seaborn==0.12.2
sentry-sdk==1.32.0
session-info==1.0.0
setproctitle==1.3.3
shtab==1.6.4
six==1.16.0
skorch==0.15.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sparse==0.14.0
stack-data==0.6.2
starlette==0.27.0
starsessions==1.3.0
statsmodels==0.14.0
stdlib-list==0.9.0
sympy==1.12
tabulate==0.9.0
tbb==2021.10.0
tensorstore==0.1.46
threadpoolctl==3.2.0
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
torch==2.0.1+cu117
torchcde==0.2.5
torchdiffeq==0.2.3
torchdyn==1.0.6
torchmetrics==1.2.0
torchsde==0.2.6
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
trampoline==0.1.2
triton==2.0.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tyro==0.5.9
tzdata==2023.3
umap-learn==0.5.4
urllib3==2.0.5
uvicorn==0.23.2
wandb==0.15.12
wcwidth==0.2.6
websocket-client==1.6.4
websockets==12.0
widgetsnbextension==4.0.9
xarray==2023.10.1
yarg==0.1.9
yarl==1.9.2
zipp==3.17.0
submitit==1.5.1
